+++ server/app.py
@@ -1,4 +1,6 @@
-from flask import Flask
+from flask import Flask, request, jsonify
+from config import DevelopmentConfig
+
 from api.ping_handler import ping_handler
 from api.home_handler import home_handler
 
@@ -7,28 +9,23 @@ from flask_sqlalchemy import SQLAlchemy
 from flask_marshmallow import Marshmallow
 from flask_script import Manager
 from flask_migrate import Migrate, MigrateCommand
-from models import shared
+
 # Models
-from models import user, appointment, meeting, availability
+from models import db, Availability, Appointment, Meeting, User
 
 
 app = Flask(__name__)
-# Load from config after
-app.config["SQLALCHEMY_DATABASE_URI"] = "sqlite:///example.sqlite"
-app.config["SQLALCHEMY_TRACK_MODIFICATIONS"] = False
+app.config.from_object(DevelopmentConfig())
 
-db = shared.db
 db.init_app(app)
-ma = Marshmallow(app)
-
-migrate = Migrate(app, db)
+migrate = Migrate(app, db, render_as_batch=True)
 manager = Manager(app)
-
 manager.add_command('db', MigrateCommand)
 
 
 app.register_blueprint(home_handler)
 app.register_blueprint(ping_handler)
+
 
 if __name__ == '__main__':
     manager.run()
